Implement claims support via new principal/identity types
