+++ Source/Csla.Shared/Csla.Shared.projitems
@@ -193,8 +193,9 @@
     <Compile Include="$(MSBuildThisFileDirectory)Rules\RuleUri.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rules\ValidationException.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)RunLocalAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Security\CslaClaimsIdentity.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Security\CslaClaimsPrincipal.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Security\CslaIdentity.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Security\CslaIdentityBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Security\CslaPrincipal.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Security\IAuthorizeReadWrite.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Security\ICheckRoles.cs" />
